Fix most of the 'comparison between signed and unsigned integer expressions' warnings

+++ compiler/lexer.cpp
@@ -488,7 +488,7 @@ namespace Aseba
 
 	bool Compiler::testNextCharacter(std::wistream &source, SourcePos &pos, wchar_t test, Token::Type tokenIfTrue)
 	{
-		if (source.peek() == int(test))
+		if ((int)source.peek() == int(test))
 		{
 			tokens.push_back(Token(tokenIfTrue, pos));
 			getNextCharacter(source, pos);
